Separate solution evaluator from particle code
